rename project to federated_aggregations
