+++ README.md
@@ -1,1 +1,1 @@
-# tff-aggregations
+# federated-aggregations
