+++ federated_aggregations/channels/channel.py
@@ -9,8 +9,8 @@ from tensorflow_federated.python.core.impl.executors import federating_executor
 from tensorflow_federated.python.core.impl.types import placement_literals
 from tf_encrypted.primitives.sodium import easy_box
 
-from tff_aggregations import utils
-from tff_aggregations.channels import key_store
+from federated_aggregations import utils
+from federated_aggregations.channels import key_store
 
 class Channel(metaclass=abc.ABCMeta):
