+++ federated_aggregations/paillier/strategy_test.py
@@ -6,10 +6,10 @@ import tensorflow as tf
 import tensorflow_federated as tff
 from tensorflow_federated.python.core.impl.executors import execution_context
 
-from tff_aggregations import channels
-from tff_aggregations.paillier import factory
-from tff_aggregations.paillier import placement
-from tff_aggregations.paillier import strategy
+from federated_aggregations import channels
+from federated_aggregations.paillier import factory
+from federated_aggregations.paillier import placement
+from federated_aggregations.paillier import strategy
 
 
 def _install_executor(executor_factory_instance):
