--- tff_aggregations/channels/__init__.py
@@ -1,4 +1,0 @@
-from tff_aggregations.channels.channel_grid import ChannelGrid
-from tff_aggregations.channels.channel import Channel
-from tff_aggregations.channels.channel import PlaintextChannel
-from tff_aggregations.channels.channel import EasyBoxChannel
